# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Store API Documentation",
        "description": "This is the API documentation for store API.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/store": {
            "get": {
                "tags": [
                    "Store"
                ],
                "summary": "Retrieve store information by postcode",
                "description": "This endpoint retrieves store information based on the provided postcode.",
                "operationId": "getStoreByPostcode",
                "parameters": [
                    {
                        "name": "postcode",
                        "in": "query",
                        "description": "Postcode to search for stores",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "eh39gu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with store data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing Bearer token"
                    },
                    "404": {
                        "description": "Store not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/add-store": {
            "post": {
                "tags": [
                    "Store"
                ],
                "summary": "Add a new store",
                "description": "This endpoint allows the user to add a new store with specific details.",
                "operationId": "addStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "The Best Food bbb"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 55.9409192
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -3.1973917
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "closeooo"
                                    },
                                    "store_type": {
                                        "type": "string",
                                        "example": "Restaurant"
                                    },
                                    "max_distance": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Store created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Store added successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or missing Bearer token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Store",
            "description": "Store"
        }
    ]
}